Split `routes.py` into a per-resource `routes/` package
Each resource now owns its own sub-router: sessions (which also carries the
node-detail and graph endpoints, both session-scoped), llm_traces, events,
middleware. The top-level ``__init__.py`` includes them under the shared
``/api/v2`` prefix.

Public surface is unchanged: ``from ..viz_api import router`` still works.
Every endpoint verified against the real event stream.
